Remove `include/grpcpp/impl/codegen/interceptor.h` (#31245)
